refactored to get rid of monolithic readStreamComponents
also removed reliance upon file path in metadata, preferring to calculate file path from `id`
