+++ src/readStream.js
@@ -3,9 +3,13 @@ var fs = require('fs');
 var batch = require('batchflow');
 var sink = require('through2-sink');
 
-var readStreamComponents = require('./readStreamComponents');
-var filterOutNamelessRecords = require('./components/filterOutNamelessRecords');
+var calculateFilePath = require('./components/calculateFilePath');
+var filterOutUnreadableFiles = require('./components/filterOutUnreadableFiles');
+var loadJSON = require('./components/loadJSON');
+var filterOutIncompleteRecords = require('./components/filterOutIncompleteRecords');
 var filterOutDeprecatedRecords = require('./components/filterOutDeprecatedRecords');
+var extractFields = require('./components/extractFields');
+var filterOutNamelessRecords = require('./components/filterOutNamelessRecords');
 
 /*
   This function finds all the `latest` files in `meta/`, CSV parses them,
@@ -13,29 +17,19 @@ var filterOutDeprecatedRecords = require('./components/filterOutDeprecatedRecord
 */
 function readData(directory, types, wofRecords, callback) {
   batch(types).parallel(2).each(function(idx, type, done) {
-    var csv_parser = parse({ delimiter: ',', columns: true });
-    var is_valid_data_file_path = readStreamComponents.is_valid_data_file_path();
-    var normalize_file_path = readStreamComponents.normalize_file_path();
-    var file_is_readable = readStreamComponents.file_is_readable(directory + 'data/');
-    var json_parse_stream = readStreamComponents.json_parse_stream(directory + 'data/');
-    var filter_incomplete_files_stream = readStreamComponents.filter_incomplete_files_stream();
-    var map_fields_stream = readStreamComponents.map_fields_stream();
-    var filter_out_nameless_records = filterOutNamelessRecords.create();
-
     fs.createReadStream(directory + 'meta/wof-' + type + '-latest.csv')
-    .pipe(csv_parser)
-    .pipe(is_valid_data_file_path)
-    .pipe(normalize_file_path)
-    .pipe(file_is_readable)
-    .pipe(json_parse_stream)
-    .pipe(filter_incomplete_files_stream)
-    .pipe(filterOutDeprecatedRecords.create())
-    .pipe(map_fields_stream)
-    .pipe(filter_out_nameless_records)
-    .pipe(sink.obj(function(wofRecord) {
-      wofRecords[wofRecord.id] = wofRecord;
-    }))
-    .on('finish', done);
+      .pipe(parse({ delimiter: ',', columns: true }))
+      .pipe(calculateFilePath.create())
+      .pipe(filterOutUnreadableFiles.create(directory + 'data/'))
+      .pipe(loadJSON.create(directory + 'data/'))
+      .pipe(filterOutIncompleteRecords.create())
+      .pipe(filterOutDeprecatedRecords.create())
+      .pipe(extractFields.create())
+      .pipe(filterOutNamelessRecords.create())
+      .pipe(sink.obj(function(wofRecord) {
+        wofRecords[wofRecord.id] = wofRecord;
+      }))
+      .on('finish', done);
 
   }).error(function(err) {
     console.error(err);
