+++ test/readStreamTest.js
@@ -16,10 +16,10 @@ tape('readStream', function(test) {
       fs.mkdirsSync('tmp/meta');
 
       // write out first meta and data files
-      fs.writeFileSync('tmp/meta/wof-type1-latest.csv', 'path\n1/2/3/4.geojson\n');
-      fs.mkdirsSync('tmp/data/1/2/3');
-      fs.writeFileSync('tmp/data/1/2/3/4.geojson', JSON.stringify({
-        id: 4,
+      fs.writeFileSync('tmp/meta/wof-type1-latest.csv', 'id\n1234567\n');
+      fs.mkdirsSync('tmp/data/123/456/7');
+      fs.writeFileSync('tmp/data/123/456/7/1234567.geojson', JSON.stringify({
+        id: 1234567,
         properties: {
           'wof:name': 'name 1',
           'wof:placetype': 'place type 1',
@@ -35,10 +35,10 @@ tape('readStream', function(test) {
       }));
 
       // write out second meta and data files
-      fs.writeFileSync('tmp/meta/wof-type2-latest.csv', 'path\n5/6/7/8.geojson\n');
-      fs.mkdirsSync('tmp/data/5/6/7');
-      fs.writeFileSync('tmp/data/5/6/7/8.geojson', JSON.stringify({
-        id: 8,
+      fs.writeFileSync('tmp/meta/wof-type2-latest.csv', 'id\n12345678\n');
+      fs.mkdirsSync('tmp/data/123/456/78');
+      fs.writeFileSync('tmp/data/123/456/78/12345678.geojson', JSON.stringify({
+        id: 12345678,
         properties: {
           'wof:name': 'name 2',
           'wof:placetype': 'place type 2',
@@ -54,10 +54,10 @@ tape('readStream', function(test) {
       // write out third meta and data files that are ignored
       // it will be ignored since 'type3' is not passed as a supported type
       // this shows that types are supported instead of all files being globbed
-      fs.writeFileSync('tmp/meta/wof-type3-latest.csv', 'path\n9/10/11/12.geojson\n');
-      fs.mkdirsSync('tmp/data/9/10/11');
-      fs.writeFileSync('tmp/data/9/10/11/12.geojson', JSON.stringify({
-        id: 12,
+      fs.writeFileSync('tmp/meta/wof-type3-latest.csv', 'id\n123456789\n');
+      fs.mkdirsSync('tmp/data/123/456/789');
+      fs.writeFileSync('tmp/data/123/456/789/123456789.geojson', JSON.stringify({
+        id: 123456789,
         properties: {
           'wof:name': 'name 3',
           'wof:placetype': 'place type 3',
@@ -82,8 +82,8 @@ tape('readStream', function(test) {
     readStream('./tmp/', ['type1', 'type2'], wofRecords, function() {
       t.equals(Object.keys(wofRecords).length, 2, 'there should be 2 records loaded');
 
-      t.deepEqual(wofRecords[4], {
-        id: 4,
+      t.deepEqual(wofRecords[1234567], {
+        id: 1234567,
         name: 'name 1',
         place_type: 'place type 1',
         parent_id: 2,
@@ -94,10 +94,10 @@ tape('readStream', function(test) {
         bounding_box: '-13.691314,49.909613,1.771169,60.847886',
         population: 98765,
         popularity: 87654
-      }, 'id 4 should have been loaded');
+      }, 'id 1234567 should have been loaded');
 
-      t.deepEqual(wofRecords[8], {
-        id: 8,
+      t.deepEqual(wofRecords[12345678], {
+        id: 12345678,
         name: 'name 2',
         place_type: 'place type 2',
         parent_id: 3,
@@ -108,7 +108,7 @@ tape('readStream', function(test) {
         bounding_box: '-24.539906,34.815009,69.033946,81.85871',
         population: undefined,
         popularity: undefined
-      }, 'id 8 should have been loaded');
+      }, 'id 12345678 should have been loaded');
 
       t.end();
