+++ test/test.js
@@ -1,8 +1,12 @@
-require ('./readStreamComponentsTest.js');
-require ('./readStreamTest.js');
+require ('./components/calculateFilePathTest.js');
+require ('./components/extractFieldsTest.js');
+require ('./components/filterOutDeprecatedRecordsTest.js');
+require ('./components/filterOutIncompleteRecordsTest.js');
+require ('./components/filterOutNamelessRecordsTest.js');
+require ('./components/filterOutUnreadableFilesTest.js');
+require ('./components/loadJSONTest.js');
+require ('./hierarchyFinderTest.js');
 require ('./importStreamTest.js');
 require ('./peliasDocGeneratorsTest.js');
+require ('./readStreamTest.js');
 require ('./wofRecordStreamTest.js');
-require ('./hierarchyFinderTest.js');
-require ('./components/filterOutNamelessRecordsTest.js');
-require ('./components/filterOutDeprecatedRecordsTest.js');
